fetching-py/requirements.txt
sanic-cors==0.9.7
python-dateutil==2.8.0
ujson==1.35
aiohttp==3.5.4
toml==0.10.0
expiringdict==1.2.1